🐛 Fixed map fails to create
Moving native object creation to the constructor has moved the native map creation before the creation of the component template, without access to the div.google-map container.
The new mechanism creates the element inside the factory.

Also, as the call to createNativeObject is done in the constructor if the base class, the map element is not yet set to the class instance and createNativeObject cannot access it. The new constructor provides a way to pass arguments to the createNativeObject method.